BS-10 : Move the folders from component to components
