+++ components/offer-table/OfferTable.tsx
